requirements.txt
#GENERAL
python-dotenv == 0.11.0

#DJANGO
django == 3.2.9
django-cors-headers == 3.10.0
django-filter == 21.1
djangorestframework == 3.12.4

psycopg2-binary == 2.9.9